+++ src/Model/ITokenMeta.ts
@@ -1,0 +1,12 @@
+interface ITokenizerResultPositions {
+	field: string
+	offsetStart: number
+	offsetEnd: number
+	position: number
+}
+
+export default interface ITokenMeta {
+	positions: ITokenizerResultPositions[]
+	weightedOccurence: number
+	magnitude: number
+}