+++ src/Model/ITokenizerResult.ts
@@ -1,0 +1,9 @@
+interface ITokenizerResultPositions {
+	offsetStart: number
+	offsetEnd: number
+	position: number
+}
+
+export default interface ITokenizerResultMap {
+	[key: string]: ITokenizerResultPositions[]
+}