+++ src/index.ts
@@ -3,5 +3,6 @@ import * as ComparisonStrategy from "./Comparison/ComparisonStrategy"
 import * as ExtractionStrategy from "./Extraction/ExtractionStrategy"
 import * as SortingStrategy from "./Sorting/SortingStrategy"
 import * as PreProcessingStrategy from "./PreProcessing/PreProcessingStrategy"
+import * as tokenizerStrategy from "./Tokenizer/TokenizerStrategy"
 
-export { Jhaystack, ComparisonStrategy, ExtractionStrategy, SortingStrategy, PreProcessingStrategy }
+export { Jhaystack, ComparisonStrategy, ExtractionStrategy, SortingStrategy, PreProcessingStrategy, tokenizerStrategy }